requirements.txt.txt
Flask
tensorflow==2.16.1 
numpy
Pillow
gunicorn